Corrected Bugs
ISSUE#10
On AtMega32U Wire.h support is not compatible with PIN Usage of DCF und
433MHZ Radios connected to PIN2 and PIN3. So i've aded a switch, that
WIRE-SUP will not be defined on a ATMega32U.

ISSUE#11
Also I've added a delay when in debug mode before first serial print

ISSUE#12
Output of Progversion was not correct. Now using the Define at the
beginning.