ATmega32U TX, Oregon Scientific, small enhancement
Enhancements to support ATMega32 Devices with TX Module PIn10 will be
used to connect the transmitter, because not all 32U devices have PIN
11.

Added a compiler option to use the FHEM module "41_Oregon", which is
already included in standard fhem distribution. Usage of Oregon Decoders
in FHEM. Code has been tested with two temperature sensors. It is also
possible to use the fhemduino_oregon Module, but this is less powerfull.

Somm corrections in oregon scientific Data output, there where some
errors with the string handling.